refactor(api): extract batch item processor into batch_item_processor.py
Move _generate_single_item_enhanced (single-item quota check, provider
selection, generation, post-processing, usage accounting, cost estimate) out of
batch.py into its own module, and drop the imports that only it used. batch.py
goes from 1265 to 1076 lines (1312 at the start of P2.3). Behavior-preserving:
imports resolve, ruff F401 back to the pre-existing baseline, full backend suite
green.
